Drive train runs from joysticks
We had trouble with speed controller groups and the differential drive object.  There were run-time errors about incorrect types.   So presently, we are just taking the raw values from the joystick and sending them to the drive motors.  This might be a bit of a hack, but seems to work for now.